+++ src/components/pages/home/Home.js
@@ -1,4 +1,4 @@
-import { Text, View } from "react-native";
+import { Text, View, StyleSheet } from "react-native";
 
 export default function Home() {
   return (
